Split resolver phase2 enum metadata tests
